# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NazLibra POS API",
        "description": "Offline-first REST API for NazLibra POS mobile application. All protected routes require Bearer token + X-Tenant-Slug header.",
        "contact": {
            "email": "support@omnevo.net"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
    "paths": {
...
        "/api/v1/sync/items": {
            "get": {
                "tags": [
                    "Sync"
                ],
                "summary": "Paginated item delta sync",
                "description": "GET /api/v1/sync/items\n\nPaginated item delta.  On first install, omit `since` to download the\nfull catalog page by page.  On subsequent calls, pass `since` from the\nprevious response's `sync_at`.\n\nQuery params:\n  since    ISO-8601 UTC timestamp (optional, inclusive >=)\n  page     integer (default 1)\n  per_page integer (default 200, max 500)",
                "operationId": "syncItems",
                "parameters": [
                    {
                        "name": "X-Tenant-Slug",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Location-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "since",
                        "in": "query",
                        "description": "ISO-8601 UTC cursor (inclusive >=)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 200,
                            "maximum": 500
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "sync_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "has_more": {
                                            "type": "boolean"
                                        },
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/sync/meta": {
            "get": {
                "tags": [
                    "Sync"
                ],
                "summary": "Sync categories, brands, units and taxes",
                "description": "GET /api/v1/sync/meta\n\nReturns all categories, brands, units and taxes for the tenant.\nThese collections are small (rarely more than a few hundred rows)\nso they are not paginated.  Pass `since` for delta.",
                "operationId": "syncMeta",
                "parameters": [
                    {
                        "name": "X-Tenant-Slug",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Location-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "since",
                        "in": "query",
                        "description": "ISO-8601 UTC cursor",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadata collections",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "sync_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "categories": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "brands": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "units": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "taxes": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/sync/contacts": {
            "get": {
                "tags": [
                    "Sync"
                ],
                "summary": "Paginated contact (customer) delta sync",
                "description": "GET /api/v1/sync/contacts\n\nPaginated customer delta.",
                "operationId": "syncContacts",
                "parameters": [
                    {
                        "name": "X-Tenant-Slug",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Location-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "since",
                        "in": "query",
                        "description": "ISO-8601 UTC cursor",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated contacts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "sync_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "has_more": {
                                            "type": "boolean"
                                        },
                                        "contacts": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}